requirements.txt
matplotlib==3.1.2
numpy==1.24.4
onnx==1.15.0
onnxsim==0.4.35
opencv_python==4.1.2.30
Pillow==8.2.0
Pillow==10.2.0
pycocotools==2.0.7
scipy==1.11.4
setuptools==68.2.2
thop==0.1.1.post2209072238
torch==1.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.60.0
